# api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Options Pricing Engine API",
    "description": "A production-ready API for option pricing and risk analytics",
    "version": "1.0.0"
  },
  "basePath": "/api",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/health": {
      "get": {
        "summary": "Get API health status",
        "responses": {
          "200": {
            "description": "Health status of all services"
          }
        }
      }
    },
    "/price/black-scholes": {
      "post": {
        "summary": "Price option using Black-Scholes formula",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "symbol": {
                  "type": "string",
                  "default": "DEMO"
                },
                "option_type": {
                  "type": "string",
                  "enum": ["call", "put"]
                },
                "strike_price": {
                  "type": "number"
                },
                "stock_price": {
                  "type": "number"
                },
                "days_to_expiration": {
                  "type": "integer"
                },
                "risk_free_rate": {
                  "type": "number",
                  "default": 0.05
                },
                "volatility": {
                  "type": "number"
                }
              },
              "required": ["strike_price", "stock_price", "days_to_expiration", "volatility"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Option price and details"
          }
        }
      }
    },
    "/price/monte-carlo": {
      "post": {
        "summary": "Price option using Monte Carlo simulation",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "symbol": {
                  "type": "string",
                  "default": "DEMO"
                },
                "option_type": {
                  "type": "string",
                  "enum": ["call", "put"]
                },
                "strike_price": {
                  "type": "number"
                },
                "stock_price": {
                  "type": "number"
                },
                "days_to_expiration": {
                  "type": "integer"
                },
                "risk_free_rate": {
                  "type": "number",
                  "default": 0.05
                },
                "volatility": {
                  "type": "number"
                },
                "num_simulations": {
                  "type": "integer",
                  "default": 10000
                }
              },
              "required": ["strike_price", "stock_price", "days_to_expiration", "volatility"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Option price with Monte Carlo statistics"
          }
        }
      }
    },
    "/greeks": {
      "post": {
        "summary": "Calculate option Greeks",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "symbol": {
                  "type": "string",
                  "default": "DEMO"
                },
                "option_type": {
                  "type": "string",
                  "enum": ["call", "put"]
                },
                "strike_price": {
                  "type": "number"
                },
                "stock_price": {
                  "type": "number"
                },
                "days_to_expiration": {
                  "type": "integer"
                },
                "risk_free_rate": {
                  "type": "number",
                  "default": 0.05
                },
                "volatility": {
                  "type": "number"
                }
              },
              "required": ["strike_price", "stock_price", "days_to_expiration", "volatility"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Option Greeks with interpretation"
          }
        }
      }
    },
    "/portfolio/analyze": {
      "post": {
        "summary": "Analyze option portfolio",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "positions": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "symbol": {
                        "type": "string"
                      },
                      "option_type": {
                        "type": "string",
                        "enum": ["call", "put"]
                      },
                      "strike_price": {
                        "type": "number"
                      },
                      "stock_price": {
                        "type": "number"
                      },
                      "days_to_expiration": {
                        "type": "integer"
                      },
                      "risk_free_rate": {
                        "type": "number"
                      },
                      "volatility": {
                        "type": "number"
                      },
                      "quantity": {
                        "type": "integer"
                      }
                    }
                  }
                }
              },
              "required": ["positions"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Portfolio analysis results"
          }
        }
      }
    }
  }
}